requirements.txt
accelerate==0.23.0
APScheduler==3.10.4
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cycler==0.11.0
dask==2023.5.0
dask-jobqueue==0.8.1
diffusers==0.21.4
distributed==2023.5.0
farmbot==1.1.0
filelock==3.12.4
Flask==2.3.2
fonttools==4.39.4
fsspec==2023.5.0
huggingface-hub==0.17.3
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
invisible-watermark==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy-loader==0.2
locket==1.0.0
MarkupSafe==2.1.1
matplotlib==3.5.3
mizani==0.9.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numpy==1.22.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
opencv-python==4.7.0.72
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.1
partd==1.4.0
patsy==0.5.3
Pillow==9.5.0
plantcv==3.14.3
plotnine==0.10.1
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.10.3
requests==2.30.0
safetensors==0.4.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
six==1.16.0
sortedcontainers==2.4.0
statsmodels==0.14.0
sympy==1.12
tblib==1.7.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.14.1
toolz==0.12.0
torch==2.1.0
tornado==6.3.2
tqdm==4.65.0
transformers==4.34.0
triton==2.1.0
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
waitress==2.1.2
webcolors==1.13
Werkzeug==2.3.3
zict==3.0.0
zipp==3.15.0
